move field type struct in a separate header
