+++ Source/Diagnostics/ComputeDiagFunctors/JFunctor.cpp
@@ -6,6 +6,7 @@
 
 #include "JFunctor.H"
 
+#include "FieldSolver/Fields.H"
 #include "Particles/MultiParticleContainer.H"
 #include "WarpX.H"
 
@@ -13,6 +14,8 @@
 #include <AMReX_Extension.H>
 #include <AMReX_IntVect.H>
 #include <AMReX_MultiFab.H>
+
+using namespace warpx::fields;
 
 JFunctor::JFunctor (const int dir, int lev,
                    amrex::IntVect crse_ratio,
