+++ Source/Diagnostics/ParticleIO.cpp
@@ -7,6 +7,7 @@
  * License: BSD-3-Clause-LBNL
  */
 
+#include "FieldSolver/Fields.H"
 #include "Particles/ParticleIO.H"
 #include "Particles/MultiParticleContainer.H"
 #include "Particles/PhysicalParticleContainer.H"
@@ -41,6 +42,7 @@
 #include <vector>
 
 using namespace amrex;
+using namespace warpx::fields;
 
 void
 LaserParticleContainer::ReadHeader (std::istream& is)
