+++ Source/Diagnostics/ReducedDiags/FieldMaximum.cpp
@@ -7,6 +7,7 @@
 
 #include "FieldMaximum.H"
 
+#include "FieldSolver/Fields.H"
 #include "Utils/TextMsg.H"
 #include "WarpX.H"
 
@@ -38,6 +39,7 @@
 #include <vector>
 
 using namespace amrex;
+using namespace warpx::fields;
 
 // constructor
 FieldMaximum::FieldMaximum (const std::string& rd_name)
