+++ Source/Diagnostics/ReducedDiags/FieldReduction.H
@@ -9,6 +9,7 @@
 #define WARPX_DIAGNOSTICS_REDUCEDDIAGS_FIELDREDUCTION_H_
 
 #include "ReducedDiags.H"
+#include "FieldSolver/Fields.H"
 #include "WarpX.H"
 
 #include <ablastr/coarsen/sample.H>
@@ -35,7 +36,6 @@
 #include <string>
 #include <type_traits>
 #include <vector>
-
 
 /**
  * This class contains a function that computes an arbitrary reduction of the fields. The function
@@ -99,15 +99,15 @@ public:
         const auto dx = geom.CellSizeArray();
 
         // get MultiFab data
-        const amrex::MultiFab & Ex = warpx.getField(FieldType::Efield_aux, lev,0);
-        const amrex::MultiFab & Ey = warpx.getField(FieldType::Efield_aux, lev,1);
-        const amrex::MultiFab & Ez = warpx.getField(FieldType::Efield_aux, lev,2);
-        const amrex::MultiFab & Bx = warpx.getField(FieldType::Bfield_aux, lev,0);
-        const amrex::MultiFab & By = warpx.getField(FieldType::Bfield_aux, lev,1);
-        const amrex::MultiFab & Bz = warpx.getField(FieldType::Bfield_aux, lev,2);
-        const amrex::MultiFab & jx = warpx.getField(FieldType::current_fp, lev,0);
-        const amrex::MultiFab & jy = warpx.getField(FieldType::current_fp, lev,1);
-        const amrex::MultiFab & jz = warpx.getField(FieldType::current_fp, lev,2);
+        const amrex::MultiFab & Ex = warpx.getField(warpx::fields::FieldType::Efield_aux, lev,0);
+        const amrex::MultiFab & Ey = warpx.getField(warpx::fields::FieldType::Efield_aux, lev,1);
+        const amrex::MultiFab & Ez = warpx.getField(warpx::fields::FieldType::Efield_aux, lev,2);
+        const amrex::MultiFab & Bx = warpx.getField(warpx::fields::FieldType::Bfield_aux, lev,0);
+        const amrex::MultiFab & By = warpx.getField(warpx::fields::FieldType::Bfield_aux, lev,1);
+        const amrex::MultiFab & Bz = warpx.getField(warpx::fields::FieldType::Bfield_aux, lev,2);
+        const amrex::MultiFab & jx = warpx.getField(warpx::fields::FieldType::current_fp, lev,0);
+        const amrex::MultiFab & jy = warpx.getField(warpx::fields::FieldType::current_fp, lev,1);
+        const amrex::MultiFab & jz = warpx.getField(warpx::fields::FieldType::current_fp, lev,2);
 
 
         // General preparation of interpolation and reduction operations
